+++ ai/saving/utils/embedder.py
@@ -1,0 +1,3 @@
+from langchain_openai import OpenAIEmbeddings
+
+embedder: OpenAIEmbeddings=OpenAIEmbeddings(model="text-embedding-3-small")
