+++ main.py
@@ -2,19 +2,20 @@ from fastapi import FastAPI, status
 import uvicorn
 import logging
 from init import init
-from ai.for_search import query_analyzer as qa
-from ai.for_search import regex_generator as rg
-from ai.for_search import tag_finder as tf
-from ai.for_search import similarity_search as ss
-from ai.for_save import query_extractor as qe
-from ai.for_save import single_adder as sa
-from ai.for_save import kakao_parser as kp
-from ai.for_save import batch_adder as ba
+
+from ai.saving.utils.embedder import embedder
+from ai.saving.adder.single_adder import single_adder
+from ai.saving.adder.batch_adder import batch_adder
+from ai.searching.regex_generator import get_regex
+from ai.searching.query_analyzer import Query_Type, query_analyzer
+from ai.searching.similarity_search import similarity_search
+from ai.searching.tag_finder import find_tag_ids
+
+from ai.saving.parser import kakao_parser as kp
 
 from models.add_memo import *
 from models.search import *
 from models.get_embedding import *
-from models.kakao_parser import *
 from models.kakao_parser import *
 
 app = FastAPI(
@@ -32,24 +33,24 @@ async def default():
 async def search(body: Arg_search):
     return_content: Res_search=Res_search()
 
-    return_content.type=qa.query_analyzer(body.content)
+    return_content.type=query_analyzer(body.content)
 
-    if return_content.type == qa.Query_Type.unspecified:
+    if return_content.type == Query_Type.unspecified:
         logging.info("[/search] unspecified query: %s", body.content)
-        return_content.type=qa.Query_Type.similarity
+        return_content.type=Query_Type.similarity
 
-    if return_content.type == qa.Query_Type.regex:
-        return_content.regex=rg.get_regex(body.content)
+    if return_content.type == Query_Type.regex:
+        return_content.regex=get_regex(body.content)
 
-    elif return_content.type == qa.Query_Type.tags:
-        return_content.tags=tf.find_tag_ids(body.content)
+    elif return_content.type == Query_Type.tags:
+        return_content.tags=find_tag_ids(body.content)
         # if the tag search result is None
         if return_content.tags == None:
             # then trying similarity search
-            return_content.type=qa.Query_Type.similarity
+            return_content.type=Query_Type.similarity
         
-    if return_content.type == qa.Query_Type.similarity:
-        return_content.processed_message, return_content.ids=ss.similarity_search(body.content)
+    if return_content.type == Query_Type.similarity:
+        return_content.processed_message, return_content.ids=similarity_search(body.content)
 
     logging.info("[/search] query: %s / query type: %s \nreturn: %s", body.content, return_content.type, return_content)
 
@@ -57,12 +58,12 @@ async def search(body: Arg_search):
 
 @app.post("/add_memo/", response_model=Res_add_memo, status_code=status.HTTP_200_OK)
 async def add_memo(body: Arg_add_memo):
-    return sa.single_adder(body)
+    return single_adder(body)
 
 @app.post("/get_embedding/", response_model=Res_get_embedding)
 async def get_embedding(body: Arg_get_embedding):
     return Res_get_embedding(
-        embedding=qe.embeddings.embed_query(body.content)
+        embedding=embedder.embed_query(body.content)
     )
 
 @app.post("/kakao-parser/", response_model=list[Res_add_memo])
@@ -72,7 +73,7 @@ async def kakao_parser(body: Arg_kakao_parser):
         Arg_add_memo(content=content, timestamp=timestamp) for content, timestamp in parsed_memolist
     ]
     
-    return await ba.batch_adder(memolist)
+    return await batch_adder(memolist)
 
 if __name__ == '__main__':
     uvicorn.run(app)
