+++ pyproject.toml
@@ -15,6 +15,7 @@ langchain-mongodb = "^0.1.6"
 pymongo = "^4.8.0"
 sentry-sdk = "^2.8.0"
 certifi = "^2024.7.4"
+langsmith = "^0.1.99"
 
 
 [build-system]
